backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
chromadb
sentence-transformers
python-dotenv
requests
beautifulsoup4
google-search-results
